Update many scripts and fix wrong month
Update many scripts and fix wrong month

+++ development/dockcross.sh
@@ -11,7 +11,7 @@ set -euo pipefail
 #//                                                          //
 #//  Script, 2021                                            //
 #//  Created: 15, July, 2021                                 //
-#//  Modified: 15, July, 2021                                //
+#//  Modified: 24, July, 2021                                //
 #//  file: -                                                 //
 #//  -                                                       //
 #//  Source: https://github.com/dockcross/dockcross                                                //
@@ -19,18 +19,22 @@ set -euo pipefail
 #//  CPU: ALL                                                //
 #//                                                          //
 #//////////////////////////////////////////////////////////////
+if (( $# == 2 )); then
+    image=$1
+    build_file=build-${image%:*}
+    shift 1
+    cmake_arg=$@
+    echo "cmake arg: $cmake_arg"
 
-image=$1
-build_file=build-${image%:*}
-shift 1
-cmake_arg=$@
-echo "cmake arg: $cmake_arg"
-
-echo "Pulling dockcross/$image"
-docker pull dockcross/$image
-echo "Make script dockcross-$image"
-docker run --rm dockcross/$image > ./dockcross-$image
-chmod +x ./dockcross-$image
-echo "Build $build_file"
-./dockcross-$image cmake -B$build_file -H. -GNinja $cmake_arg
-./dockcross-$image ninja -C$build_file
+    echo "Pulling dockcross/$image"
+    docker pull dockcross/$image
+    echo "Make script dockcross-$image"
+    docker run --rm dockcross/$image > ./dockcross-$image
+    chmod +x ./dockcross-$image
+    echo "Build $build_file"
+    ./dockcross-$image cmake -B$build_file -H. -GNinja $cmake_arg
+    ./dockcross-$image ninja -C$build_file
+else
+    echo "Usage: ${0##*/} <docker imag (ex: linux-x64/linux-arm64...)> <cmake arg.>"
+    exit 1
+fi
